use new assertion for all migration/create tests
