+++ tests/framework/console/controllers/migrate_create/default.php
@@ -1,4 +1,5 @@
 <?php
+
 return <<<CODE
 <?php
