+++ tests/framework/console/controllers/migrate_create/drop_fields.php
@@ -1,0 +1,34 @@
+<?php
+
+return <<<CODE
+<?php
+
+use yii\db\Migration;
+
+/**
+ * Handles the dropping for table `test`.
+ */
+class {$class} extends Migration
+{
+    /**
+     * @inheritdoc
+     */
+    public function up()
+    {
+        \$this->dropTable('test');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function down()
+    {
+        \$this->createTable('test', [
+            'id' => \$this->primaryKey(),
+            'body' => \$this->text()->notNull(),
+            'price' => \$this->money(11,2),
+        ]);
+    }
+}
+
+CODE;
